+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 2.1.0 [in progress]
+## 2.1.0 [2020-10-02]
 ### Features
 1. [#193](https://github.com/influxdata/influxdb-client-go/pull/193) Added authentication using username and password. See `UsersAPI.SignIn()` and `UsersAPI.SignOut()`